[skip ci] :green_heart: Support API docs for v4
- Move v4 docs to v4 dir
- Use API docs workflow from ci repo
