+++ v4/docs/doxygen.conf
@@ -1,6 +1,6 @@
 # Adapted from https://github.com/xtensor-stack/xtensor/blob/master/docs/Doxyfile
 PROJECT_NAME = "libhal"
-INPUT = v4/include
+INPUT = ../include
 FILE_PATTERNS = *.hpp
 EXTRACT_STATIC = YES
 EXCLUDE_PATTERNS = */third_party/*
