+++ v4/docs/libhal/accelerometer.md
