+++ v4/docs/libhal/adc.md
