+++ v4/docs/libhal/angular_velocity_sensor.md
