+++ v4/docs/libhal/dac.md
