+++ v4/docs/libhal/error.md
