+++ v4/docs/libhal/functional.md
