+++ v4/docs/libhal/gyroscope.md
