+++ v4/docs/libhal/i2c.md
