+++ v4/docs/libhal/index.rst
