+++ v4/docs/libhal/input_pin.md
