+++ v4/docs/libhal/magnetometer.md
