+++ v4/docs/libhal/motor.md
