+++ v4/docs/libhal/output_pin.md
