+++ v4/docs/libhal/rotation_sensor.md
