+++ v4/docs/libhal/spi.md
