+++ v4/docs/libhal/temperature_sensor.md
