+++ v4/docs/libhal/timer.md
